+++ packages/admin/client/views/pages/places.js
@@ -30,7 +30,9 @@ Template.itemPlaceAdmin_admin_btns.onRendered(function() {
 	var self = this;
 	self.$('.item-btn-del').btsConfirmButton(function(e) {
 		e.stopPropagation();
-		
+
+		K.Map.removeItem(self.data);
+		console.log(self.data)
 		K.Admin.removePlace(self.data.id);
 		delete K.placesById[self.data.id];