+++ packages/osm/client/views/helpers.js
@@ -1,0 +1,6 @@
+
+
+Template.registerHelper('osmEditUrlById', function(osmid) {
+	var id = osmid.split('/').join('=');
+	return 'https://www.openstreetmap.org/edit?'+id+'&amp;editor=id';
+});